requirements.txt
aiohttp==3.9.1
pandas==2.1.4
pyodbc==5.0.1
pyyaml==6.0.1
tenacity==8.2.3
tqdm==4.66.1
python-dateutil==2.8.2
requests==2.31.0